chore: 쿠폰에 대한 타입 정의
squash
